+++ examples/demos/infrastructure_provisioning_without_i3s.tf
@@ -1,4 +1,4 @@
-/*   Infrastructure Provisioning without Storage and Image Stramear
+/*   Infrastructure Provisioning with Networks and Storage
 
 	Be able to provision compute (with server settings), networking, and storage.
 	Create a server profile template with the following options:
@@ -10,15 +10,15 @@
 */
 
 provider "oneview" {
-	ov_username = "<ov_username>"
-	ov_password = "<ov_password>"
-	ov_endpoint = "<ov_endpoint>"
-	ov_sslverify = false
-	ov_apiversion = <ov_apiversion>
-	ov_ifmatch = "*"
+  ov_username =   "${var.username}"
+  ov_password =   "${var.password}"
+  ov_endpoint =   "${var.endpoint}"
+  ov_sslverify =  "${var.ssl_enabled}"
+  ov_apiversion = 2200
+  ov_ifmatch = "*"
 }
 
-//Create Networks
+# Create Networks
 resource "oneview_ethernet_network" "ethernetnetwork" {
   name = "UCENet"
   type = "ethernet-networkV4"
@@ -30,9 +30,10 @@ resource "oneview_fc_network" "fc_network" {
   type = "fc-networkV4"
 }
 
+# Create Logical Interconnect Group
 resource "oneview_logical_interconnect_group" "logical_interconnect_group" {
   name = "UCLIG"
-  type = "logical-interconnect-groupV5"
+  type = "logical-interconnect-groupV8"
   interconnect_bay_set = 3
   redundancy_type = "HighlyAvailable"
   enclosure_indexes = [1, 2, 3]
@@ -70,9 +71,27 @@ resource "oneview_logical_interconnect_group" "logical_interconnect_group" {
     },
 
   ]
+  uplink_set = [
+    {
+      network_type = "Ethernet"
+      ethernet_network_type = "Tagged"
+      name = "test"
+      logical_port_config = [
+        {
+          desired_speed = "Auto"
+          port_num = [66]
+          bay_num = 3
+          enclosure_num = 1
+          primary_port = false
+        },
+      ]
+      mode = "Auto"
+      network_uris = ["${oneview_ethernet_network.ethernetnetwork.uri}"]
+    }
+  ]
 }
 
-// Create enclosure group
+# Create Enclosure Group
 resource "oneview_enclosure_group" "enclosure_group" {
   name = "UCEG"
   ip_addressing_mode = "DHCP"
@@ -88,37 +107,47 @@ resource "oneview_enclosure_group" "enclosure_group" {
       logical_interconnect_group_uri = "${oneview_logical_interconnect_group.logical_interconnect_group.uri}"
     },
   ]
+  depends_on = ["oneview_logical_interconnect_group.logical_interconnect_group"]
 }
 
+# Create Logical Enclosure
 resource "oneview_logical_enclosure" "logical_enclosure" {
   name = "UCLE"
   enclosure_uris = ["/rest/enclosures/0000000000A66101","/rest/enclosures/0000000000A66102","/rest/enclosures/0000000000A66103"]
   enclosure_group_uri = "${oneview_enclosure_group.enclosure_group.uri}"
+  depends_on = [ "oneview_enclosure_group.enclosure_group"]
 }
 
-// Get Server Hardware Type
+# Get Server Hardware Type
 data "oneview_server_hardware_type" "server_hardware_type" {
 	name = "SY 480 Gen9 1"
 }
 
-// Get Server Hardware 
+# Get Server Hardware 
 data "oneview_server_hardware" "server_hardware" {
   name = "0000A66101, bay 7"
 }
 
-// Create Server Profile Template with Network and BOOT
+# Create Server Profile Template with Network and BOOT
 resource "oneview_server_profile_template" "server_profile_template" {
-	name = "UCSPT"
-	type = "ServerProfileTemplateV5"
+	name = "UCSPT-US"
+	type = "ServerProfileTemplateV8"
 	enclosure_group = "${oneview_enclosure_group.enclosure_group.name}"
 	server_hardware_type = "${data.oneview_server_hardware_type.server_hardware_type.name}"
   network = [{
-    name = "UcEthNetConn"
+    name = "UcFCNetConn"
     function_type = "FibreChannel"
     port_id = "Mezz 2:1"
     network_uri = "${oneview_fc_network.fc_network.uri}"
     requested_mbps = ""
-  }]
+  },
+  { name = "UcEthNetConn"
+    function_type = "Ethernet"
+    port_id = "Mezz 3:1-a"
+    network_uri = "${oneview_fc_network.fc_network.uri}"
+    requested_mbps = ""
+  },
+]
   manage_connections = true
   boot_order = ["HardDisk"]
   boot_mode = {
@@ -129,11 +158,11 @@ resource "oneview_server_profile_template" "server_profile_template" {
   depends_on = ["oneview_logical_enclosure.logical_enclosure"]
 }
 
-// Create Server Profile with Network and BOOT
+# Create Server Profile with Network and BOOT
 resource "oneview_server_profile" "server_profile" {
   name = "UCSP"
   template = "${oneview_server_profile_template.server_profile_template.name}"
   hardware_name = "${data.oneview_server_hardware.server_hardware.name}"
-  type = "ServerProfileV9"
+  type = "ServerProfileV12"
   depends_on = ["oneview_logical_enclosure.logical_enclosure"]
 }