+++ CHANGELOG.md
@@ -4,8 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 # [v1.6.0](Unreleased)
 ### Notes
 - This release supports API2200 minimally where we can use OneView API2200 with this SDK. No new fields are added/deleted to support API2200.
+  This release adds variable.tf file to provide easy access for adding Appliance Credentials through environment variables or during runtime. 
 
 ### Oneview Features supported
+- Deployment Plans
 - Enclosure
 - Enclosure Group
 - Ethernet Network